chg: move the definition of ListOrObject and the make_list function from button to widgets as it can be used by multiple composite widgets
